Fix install scripts for RPM
Without this change, the npm-cli.js script is configured
with a hardcoded hash-bang path to a copy of node that we
build. This wouldn't be a problem except that cpack (or
rpm-build?) detects the hash-bang path as being a dependency
and complains at install time.

This change adds a bin/npm script that works regardless of
whether/where shiny-server is installed, and rewrites the
hash-bang (using the npm package's own relocate.sh script)
to use "/usr/bin/env node".