Refactored local config check to separate router.
In order to properly cache the local config files (which is important for perdormance), we need to be aware of the restart timestamp, which means the router handling the local config checks must sit outside of the RestartRouter. Since our config router is the terminal, inner-most router, that wasn't possible. This moves the local config check to a separate router which can sit outside of all others.