Launch shiny-server through shim binary
This replaces the bin/shiny-server bash script we were previously using.
The problem with the former was that it caused at least one fork, which
confused upstart (I tried using "expect" directives to get it to detect
the right PID but the obvious incantations were not working).

Using the new bin/shiny-server binary, we can optionally launch
daemonized by using --daemonize, or just run direct which does not fork
at all.

Note that the upstart/init.d scripts may need to be updated to take full
advantage.