Pass RSTUDIO_PANDOC through stdin.
CentOS 6.3 seems to wipe it when going through bash --login, so we need another way to pass it in. We're actually not clear whether or not it wipes LANG and family, but if it does we expect it will provide us with a reasonable system default, so that's less of a concern.