Add systemd script for Ubuntu 15.04
The systemd script could probably be a lot simpler; we shouldn't
need a PIDFILE if we're not running in daemon mode. And I have a
feeling we shouldn't be writing into /var/log/shiny-server.log for
the main server, but rather to the systemd journal.

However, in the interest of getting a working build out to 15.04
users ASAP, and not having to retest Red Hat 7, we'll just take
our existing systemd file with minimal changes.