Copy instead of move Pandoc
By keeping the files that yum installs present, we make yum happier when
it goes to uninstall those files. So we'll copy instead of mv as a quick
fix.

Eventually, the right answer is actually to move this intelligence to
CMake and only ship with the appropriate version of pandoc. I had spent
a few hours on it many months ago and ended up deciding this was better
than wasting a day on it.