Explicitly cd into the appropriate directory.
On (at least) RHEL7, su's -p switch is clobbered by the --login switch, meaning that the current directory is no longer maintained across the su call. Instead, the user's home directory will be the dir of the executed command. For us, that means the app is running out of the wrong directory and will not properly respect things like an .RProfile in the app dir.