Don't forbid storage of templated files.
no-cache is sufficient to discourage browsers from serving the
cached version of a file. With browser history and/or tab
reopening, however, the browser /will/ use the cached version of the
file (tested on Chrome and Firefox). Thus, if a file is essential in
order to reload the state of a tab (as is the case for the templated
shiny-server.js), you need to allow the browser to store so that it
will have a copy of shiny-server.js when it re-opens the tab.

This commit restores the "disconnect" dialog box when reopening a closed
tab while offline.