Remove node_modules
Stop vendoring all of our dependencies during development. The only
downside of this is if dependencies get unpublished or npmjs.org
and/or github are down when we try to build on jenkins. If that is
a problem we can use something like "slc build" to fully vendor
dependencies in a deployment branch.

Use "npm shrinkwrap" to snapshot the exact dependency versions.

The "i" package was manually changed to 0.3.3 because 0.3.2 was
accidentally unpublished by the author and can't be installed
anymore.

The make-package.sh script now does npm install, not just rebuild.
I'm not sure rebuild is necessary (probably not) but leaving it
for now.