Upgrade shiny-server-client
This upgrade fixes a regression where the protocol chooser UI
shows all protocols, regardless of whether the server has them
enabled or not.

Older versions of SS/SSP hide the protocols that are disabled.
Versions that used prior versions of shiny-server-client show
all the protocols. This change still shows all the protocols,
but now greys out the ones that are not enabled by the server.