Fix #506: R process not terminating in certain scenario
Bitten again by node-http-proxy event changes :( When an ECONNRESET
occurs between the proxy and the target (Shiny process), we were
expecting to see that error come in on the "error" event (and in
earlier versions it was "proxyError"). But earlier this year, this
specific error code was pulled out of the "error" event and raised
as a unique "econnreset" event, without the docs being updated.

Also added a test so the build will fail if we upgrade http-proxy
and it adds new events.