Fix SockJS transport fallback
Before this fix, some proxy configurations were working for Shiny
Server 1.4.3 but not in 1.5.0. (I can't vouch for the state of
versions in between.)

Basically, with 1.4.3 SockJS was able to fall gracefully back from
websockets, to xhr-streaming, to xhr-polling. With 1.5.0, SockJS
does all this but when it finally finds a transport that works,
Shiny Server kills the connection, and in the server log there's a
message about RobustSockJS collision.

I believe this is because the proxy is letting the request through
correctly, but interrupting the response (for xhr-streaming). So
the server thinks the conversation has started, but the client
doesn't. When the client tries the next transport, it's still
trying to start a conversation, but now the server thinks the
client is using a robust-id (i.e. conversation ID) for a convo
that's already in progress, hence, collision.

This fix introduces a "nascent" flag for robust connections that
means the connection is newly created. Connections are considered
nascent only until they receive their first message from the client.
A nascent connection can be connected to using either "new" or
"continue" semantics (whereas before, any existing connection could
only be continued; attempting to treat them as new would result in
the RobustSockJS collision warning, and immediate closure of the
connection).

It is still NOT allowed to connect to an existing, non-nascent
connection using "new"--in that case the client and server are in
truly inconsistent states.

To repro this, you can use an nginx proxy configuration like so:

location / {
  proxy_pass http://localhost:3838;
  proxy_redirect http://localhost:3838/ $scheme://$host/;
}

and make sure xhr-streaming is enabled.

To see a correctly configured nginx proxy, visit:
https://support.rstudio.com/hc/en-us/articles/213733868-Running-Shiny-Server-with-a-Proxy