Refactor of scheduler
This refactor is intended to simplify the scheduler
logic; specifically, to eliminate a gap in resource
acquisition tracking when a worker is starting up.

Previously, worker clients like lib/proxy/http.js and
lib/proxy/sockjs.js would call getWorker_p, then
after the returned promise is successfully resolved
with an app worker handle, call acquire() on that
handle. The resolution of the promise doesn’t
happen until the app has successfully started up
(including global.R executing) so it may take a long
time. During this window of time, the scheduler has
effectively handed out the resource (the worker)
but the bookkeeping has not been done yet, since
acquire() is not called until later. This gap between
resource assignment and bookkeeping is a problem
for enforcing resource limits. To mitigate this, there
was a `pendingConn` connection count that would
track the number of http/sock connections that are
expected to happen after startup completes.

This commit modifies the implementation of
getWorker_p (and associated underlying functions)
to return not a promise to the worker handle, but
instead, a new abstraction called a WorkerEntry.
This WorkerEntry is returned synchronously and
has acquire() and release() methods, as well as a
way to retrieve the promise to the worker handle.
Callers should call acquire() immediately after
getWorker returns, eliminating the bookkeeping
gap and thus obsoleting pendingConn.