Re-introduce pendingConn; cleaner release
Until it was removed, pendingConn counted the number of acquire()
calls that were expected but not yet performed; this was because
getWorker_p() and acquire() calls were separated by arbitrary
amounts of time if a worker was in the process of starting up.
That window was removed recently, as acquire() is now called right
after getWorker().

Now we need to bring back pendingConn to mean something else: when
an app page (root of a Shiny app, i.e. "/", or a Shiny doc, i.e.
"/whatever.Rmd") is called, it's usually safe to assume that a new
session is about to begin. However, there really is an inherent
delay between 1) the app page being requested, 2) the app page
being returned, 3) the sockjs connection actually arriving at the
server. So pendingConn closes the window between 1 and 3.

I also refactored the logic that caused release() to be called. I
didn't like assigning parent-scope variables, it's hard to follow
and feels error prone.