Set up logging earlier in R subprocess
I recently spent some time chasing a [silly] mis-configuration that
resulted in the R subprocess not starting up.  I wasn't getting any
useful log messages because it was dying before logging had been set up.

This change moves the logging setup code as early as possible in the
setup processes so that errors are logged and usable.