Fix "Can't set headers after they are sent" error in logs
Repro case: Navigate to a Shiny app URL but leave off the trailing /.

The piece of code that detects that is supposed to send a redirect
and then return true; to indicate that the request was fully handled;
but instead, it sends the redirect and then return false;, which
means that it could not be handled and the next handler should take
a crack at it. The final handler in the chain sends a 404 (i.e.
that's the default behavior if no handlers know what to do with it),
but that fails because the redirect was already sent.