Fix #637 (SSP): Can't remove headers after they are sent
We used to go through some trouble to have HTTP keepalive between
browsers and Shiny Server, but not keepalive (Connection: close)
between Shiny Server and Shiny (R process). Previously we cared
about this because when we used unix domain sockets to communicate
between Shiny Server and Shiny, connection pooling didn't work
right (at least not in those days, Node v0.8 probably). Now we
don't use unix domain sockets, but we found out during Winston's
httpuv background-thread work that with keepalive enabled, we
weren't properly resetting header information between requests.
This isn't so bad in local cases, but for the SS->Shiny connection
this could be very bad since we could be proxying requests on
behalf of different users over the same connections.

It turns out that we didn't need to do this work anyway; it is
the default behavior of node-http-proxy anyway:
https://github.com/nodejitsu/node-http-proxy/blob/812757541db6a6eb9021dc459830d7c0a9b4429d/lib/http-proxy/common.js#L66-L75

I've confirmed using wireshark that indeed the TCP connections
between SS and Shiny are closed every time, but kept alive
between browser and SS.

Note that neither I (nor, to my knowledge, Alan or Fereshteh)
were ever able to repro the issue, which was supposed to cause
this crash when a POST request was proxied.