Move server (re)starting logic to posttrans, for correct upgrades
During upgrades, server restarts were happening during a phase of the rpm
lifecycle where files from BOTH old and new versions of the package were
on disk. In some cases we can get the wrong behavior (e.g. npm packages
existing in a nested location in the old version, whereas it moves to the
top level in the new version; we would (and did) pick up the old version
until the server is restarted).

The correct place to restart is during posttrans. CPackRPM doesn't have
a posttrans feature, but we can fake it by tacking %posttrans to the end
of any other script.