Update install-node.sh: NODE_FILENAME
When I tried to install the node through this script on my raspberry, independent of the chosen NODE_SHA256 the downloaded node was the same because of the NODE_FILENAME definition, that was hold on linux-x64. With this change, it´s more explicit the need for select the right distribution - This change has revelance once the number of unsucess on installing shiny server on raspberry is high, in general because of this issue.