Fix build
This was due to an update to shiny-server-client that I made, which
was only intended to upgrade a dependency whose pinned version had
been unpublished. I accidentally commited to shiny-server-client's
npm-shrinkwrap.json all of the devDependencies as well. That caused
shiny-server to build with all of those dependencies, and some of
those dependencies do not have licenses listed, which caused our
tools/preflight.sh script to complain.

I've now changed shiny-server-client to use package-lock.json, which
will make it harder to get into this situation in the future.