Fix persistent launch error
In lib/scheduler/scheduler.js there's a call to posix.getpwnam. If
an error is raised there, cleanup is incomplete; in particular, the
workerEntry for the new (now failed) worker is not removed. This
broken worker is then reused as if it is working, returning the
same error to every new client.

This fix simply removes the workerEntry.