SingleAppRouter.getAppSpec_p does too much work before matching URL
A customer reported that a broken `app_dir` (e.g. root directory
that doesn't exist) caused many apps on the server to be broken,
not just the broken app itself. This is because during routing
(i.e. getAppSpec_p()), the locations are tried one after another
to see if they match the request URL. In the case of app_dir,
the SingleAppRouter.getAppSpec_p method was doing a bunch of
directory reading before even looking to see if the URL was
a match; if that directory reading failed, the request would
see a 500 error or whatever.

With this change, an error should only be seen when a broken
app is actually being requested.