updating the way jenkins stores artifacts
adding new Jenkinsfile.internal for testing while we transition all
branches to the new Jenkins job.

save the built packages as artifacts in jenkins
commented out uploading to the s3 bucket for now
upload to the the cloudsmith repositories

removed the external ssos test cases jenkins job because it is
out of date.